async-model: force async models to be async, move test
